requirements.txt

pypdf2
psycopg2-binary
python-dotenv
openai